requirements.txt
django~=5.1
djangorestframework~=3.15.2
django-oauth-toolkit~=2.4.0

requests~=2.32.3
factory_boy~=3.3.1
pytest
pytest-django
pytest-mock
